requirements.txt
# Core dependencies
torch==2.2.0
torchvision==0.17.0
pandas
Pillow

# Development
pytest
